move src/App.* into src/containers/
